[qa] Design Fix
- feature
