+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/MyPageArgument.kt
@@ -23,3 +23,4 @@ sealed interface MyPageEvent
 sealed interface MyPageIntent{
     data object RefreshData : MyPageIntent
 }
+
