+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/others/account/withdrawal/WithdrawalConstant.kt
@@ -1,0 +1,7 @@
+package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.account.withdrawal
+
+import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.MyPageConstant
+
+object WithdrawalConstant {
+    const val ROUTE: String = "${MyPageConstant.ROUTE}/withdrawal"
+}
