+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/others/account/withdrawal/WithdrawalData.kt
@@ -1,4 +1,4 @@
-package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.logout.withdrawal
+package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.account.withdrawal
 
 import androidx.compose.runtime.Immutable
 import kr.linkerbell.campusmarket.android.domain.model.nonfeature.user.MyProfile
