+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/others/account/withdrawal/WithdrawalDestination.kt
@@ -1,25 +1,33 @@
-package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.logout
+package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.account.withdrawal
 
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavType
 import androidx.navigation.compose.composable
+import androidx.navigation.navArgument
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.ErrorObserver
 
-fun NavGraphBuilder.logoutDestination(
+fun NavGraphBuilder.withdrawalDestination(
     navController: NavController
 ) {
     composable(
-        route = LogoutConstant.ROUTE
+        route = WithdrawalConstant.ROUTE,
+        arguments = listOf(
+            navArgument(WithdrawalConstant.ROUTE) {
+                type = NavType.StringType
+                defaultValue = ""
+            }
+        )
     ) {
-        val viewModel: LogoutViewModel = hiltViewModel()
+        val viewModel: WithdrawalViewModel = hiltViewModel()
 
-        val argument: LogoutArgument = let {
+        val argument: WithdrawalArgument = let {
             val state by viewModel.state.collectAsStateWithLifecycle()
 
-            LogoutArgument(
+            WithdrawalArgument(
                 state = state,
                 event = viewModel.event,
                 intent = viewModel::onIntent,
@@ -28,10 +36,19 @@ fun NavGraphBuilder.logoutDestination(
             )
         }
 
+        val data: WithdrawalData = let {
+            val myProfile by viewModel.myProfile.collectAsStateWithLifecycle()
+
+            WithdrawalData(
+                myProfile = myProfile
+            )
+        }
+
         ErrorObserver(viewModel)
-        LogoutScreen(
+        WithdrawalScreen(
             navController = navController,
             argument = argument,
+            data = data
         )
     }
 }
