--- presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/others/logout/withdrawal/WithdrawalConstant.kt
@@ -1,7 +1,0 @@
-package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.logout.withdrawal
-
-import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.logout.LogoutConstant
-
-object WithdrawalConstant {
-    const val ROUTE: String = "${LogoutConstant.ROUTE}/withdrawal"
-}
