+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/others/userprofile/recent/review/RecentReviewScreen.kt
@@ -1,9 +1,6 @@
 package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.others.userprofile.recent.review
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -16,19 +13,14 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
@@ -46,20 +38,14 @@ import kotlinx.datetime.toLocalDateTime
 import kr.linkerbell.campusmarket.android.common.util.coroutine.event.MutableEventFlow
 import kr.linkerbell.campusmarket.android.domain.model.feature.mypage.UserReview
 import kr.linkerbell.campusmarket.android.presentation.R
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Black
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Blue400
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Body1
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Body2
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Caption2
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray200
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray600
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray900
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Headline2
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Space56
 import kr.linkerbell.campusmarket.android.presentation.common.theme.White
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.safeNavigateUp
 import kr.linkerbell.campusmarket.android.presentation.common.view.RippleBox
-import kr.linkerbell.campusmarket.android.presentation.common.view.image.PostImage
+import kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.common.ReviewCard
 
 @Composable
 fun RecentReviewScreen(
@@ -105,7 +91,7 @@ fun RecentReviewScreen(
                 )
             }
             Text(
-                text = "최근 평가",
+                text = "최근 리뷰 목록",
                 style = Headline2
             )
         }
@@ -136,7 +122,7 @@ fun RecentReviewScreen(
                 }
             }
             LazyColumn(
-                contentPadding = PaddingValues(vertical = 16.dp, horizontal = 20.dp)
+                contentPadding = PaddingValues(vertical = 16.dp, horizontal = 16.dp)
             ) {
                 items(
                     count = recentReviewList.itemCount,
@@ -149,118 +135,9 @@ fun RecentReviewScreen(
                 ) { index ->
                     val review = recentReviewList[index] ?: return@items
                     ReviewCard(review)
-                    Spacer(modifier = Modifier.height(8.dp))
+                    Spacer(modifier = Modifier.padding(8.dp))
                 }
             }
-        }
-    }
-}
-
-@Composable
-private fun ReviewCard(review: UserReview) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(8.dp))
-            .background(Gray200)
-            .padding(8.dp)
-    ) {
-        Row {
-            PostImage(
-                data = review.profileImage,
-                modifier = Modifier
-                    .size(40.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(
-                        color = White,
-                        shape = RoundedCornerShape(4.dp)
-                    )
-                    .border(
-                        1.dp,
-                        Gray900,
-                        shape = RoundedCornerShape(12.dp)
-                    )
-            )
-            Column(
-                modifier = Modifier.padding(start = 8.dp),
-                verticalArrangement = Arrangement.SpaceEvenly
-            ) {
-                Row(
-                    modifier = Modifier.padding(bottom = 4.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = review.nickname,
-                        style = Headline2,
-                        color = Black
-                    )
-                    Text(
-                        text = review.createdAt.date.toString(),
-                        style = Body2,
-                        color = Gray600,
-                        modifier = Modifier.padding(start = 4.dp)
-                    )
-                }
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    RatingStars(
-                        rating = review.rating,
-                        starSize = 14.dp,
-                        interval = 2.dp
-                    )
-                    Text(
-                        text = "(${review.rating})",
-                        style = Body2,
-                        color = Black,
-                        modifier = Modifier.padding(start = 4.dp)
-                    )
-                }
-            }
-        }
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(vertical = 4.dp)
-        ) {
-            Text(
-                text = review.description,
-                overflow = TextOverflow.Ellipsis,
-                maxLines = 4,
-                style = Body1,
-                color = Gray900,
-                modifier = Modifier.padding(4.dp)
-            )
-        }
-    }
-}
-
-@Composable
-private fun RatingStars(
-    rating: Int = 0,
-    starSize: Dp = 20.dp,
-    interval: Dp = 4.dp
-) {
-    val adjustedRating = rating.coerceIn(0, 10)
-
-    @Composable
-    fun StarImage(resourceId: Int, size: Dp) {
-        Image(
-            painter = painterResource(id = resourceId),
-            contentDescription = null,
-            colorFilter = ColorFilter.tint(Blue400),
-            modifier = Modifier.size(size)
-        )
-    }
-
-    Row {
-        repeat(5) { index ->
-            val starPoints = (index + 1) * 2
-            val resource = when {
-                adjustedRating >= starPoints -> R.drawable.filled_star
-                adjustedRating >= starPoints - 1 -> R.drawable.half_filled_star
-                else -> R.drawable.empty_star
-            }
-            StarImage(resource, starSize)
-            Spacer(modifier = Modifier.padding(end = interval))
         }
     }
 }
@@ -295,7 +172,7 @@ private fun RecentReviewScreenPreview() {
                             profileImage = "",
                             description = "아주 좋아요",
                             rating = 10,
-                            createdAt = LocalDateTime(2024, 11, 22, 15, 30, 0)
+                            createdAt = LocalDateTime(2024, 11, 23, 15, 30, 0)
                         )
                     )
                 )
