+++ presentation/src/main/kotlin/kr/linkerbell/campusmarket/android/presentation/ui/main/home/mypage/recent/likes/MyLikesScreen.kt
@@ -1,24 +1,15 @@
 package kr.linkerbell.campusmarket.android.presentation.ui.main.home.mypage.recent.likes
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.filled.FavoriteBorder
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -28,12 +19,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Color.Companion.LightGray
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
@@ -41,7 +27,6 @@ import androidx.constraintlayout.compose.Dimension
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import androidx.paging.PagingData
-import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -49,21 +34,16 @@ import kotlinx.coroutines.plus
 import kr.linkerbell.campusmarket.android.common.util.coroutine.event.MutableEventFlow
 import kr.linkerbell.campusmarket.android.domain.model.feature.trade.SummarizedTrade
 import kr.linkerbell.campusmarket.android.presentation.R
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Blue200
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Blue400
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Body1
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Caption2
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray200
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Gray600
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Headline2
-import kr.linkerbell.campusmarket.android.presentation.common.theme.Headline3
 import kr.linkerbell.campusmarket.android.presentation.common.theme.Space56
 import kr.linkerbell.campusmarket.android.presentation.common.theme.White
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.isEmpty
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.makeRoute
 import kr.linkerbell.campusmarket.android.presentation.common.util.compose.safeNavigateUp
 import kr.linkerbell.campusmarket.android.presentation.common.view.RippleBox
-import kr.linkerbell.campusmarket.android.presentation.common.view.image.PostImage
+import kr.linkerbell.campusmarket.android.presentation.ui.main.home.trade.common.SummarizedTradeCard
 import kr.linkerbell.campusmarket.android.presentation.ui.main.home.trade.info.TradeInfoConstant
 
 @Composable
@@ -126,7 +106,7 @@ fun MyLikesScreen(
         }
         Column(
             modifier = Modifier
-                .padding(vertical = 8.dp)
+                .padding(vertical = 16.dp, horizontal = 16.dp)
                 .constrainAs(contents) {
                     top.linkTo(topBar.bottom)
                     start.linkTo(parent.start)
@@ -136,154 +116,34 @@ fun MyLikesScreen(
                     height = Dimension.fillToConstraints
                 }
         ) {
-            MyRecentTradeListScreen(
-                myLikes = myLikes,
-                onItemClicked = { itemId ->
-                    onItemClicked(itemId)
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun MyRecentTradeListScreen(
-    myLikes: LazyPagingItems<SummarizedTrade>,
-    onItemClicked: (Long) -> Unit
-) {
-    if (myLikes.isEmpty()) {
-        Box(
-            modifier = Modifier.fillMaxWidth(),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "보여줄 항목이 없어요",
-                style = Caption2,
-                color = Gray600,
-                modifier = Modifier.padding(start = 8.dp, top = 8.dp)
-            )
-        }
-    }
-    HorizontalDivider(
-        thickness = 1.dp,
-        color = Gray200,
-        modifier = Modifier.padding(horizontal = 2.dp)
-    )
-    LazyColumn {
-        items(
-            count = myLikes.itemCount,
-            key = { index -> myLikes[index]?.itemId ?: -1 }
-        ) { index ->
-            val trade = myLikes[index] ?: return@items
-            TradeItemCard(
-                item = trade,
-                onItemCardClicked = {
-                    onItemClicked(trade.itemId)
-                }
-            )
-        }
-    }
-}
-
-@Composable
-private fun TradeItemCard(
-    item: SummarizedTrade,
-    onItemCardClicked: (Long) -> Unit
-) {
-    Box(
-        Modifier
-            .shadow(4.dp)
-            .clip(RoundedCornerShape(5.dp))
-            .clickable {
-                onItemCardClicked(item.itemId)
-            }
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.White)
-                .padding(8.dp),
-        ) {
-            PostImage(
-                data = item.thumbnail,
-                modifier = Modifier
-                    .size(100.dp)
-                    .clip(RoundedCornerShape(8.dp))
-            )
-            Column(
-                modifier = Modifier.padding(start = 10.dp)
-            ) {
-                Text(
-                    text = item.title,
-                    style = Headline3,
-                    overflow = TextOverflow.Ellipsis,
-                    maxLines = 1,
-                    modifier = Modifier.padding(bottom = 4.dp),
-                )
-
-                Text(
-                    text = item.nickname,
-                    style = Caption2,
-                    modifier = Modifier.padding(bottom = 4.dp)
-                )
-                Text(
-                    text = "${item.chatCount} 명이 대화중",
-                    style = Caption2,
-                    modifier = Modifier.padding(bottom = 4.dp)
-                )
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .widthIn(min = 100.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween
+            if (myLikes.isEmpty()) {
+                Box(
+                    modifier = Modifier.fillMaxWidth(),
+                    contentAlignment = Alignment.Center
                 ) {
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        TradeItemStatus(isSold = item.itemStatus == "SOLDOUT")
-                        Text(
-                            text = "${item.price} 원",
-                            overflow = TextOverflow.Ellipsis,
-                            modifier = Modifier.padding(4.dp)
-                        )
-                    }
-                    Row(
-                        modifier = Modifier.padding(start = 8.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        val favIcon =
-                            if (item.isLiked) Icons.Default.Favorite else Icons.Default.FavoriteBorder
-                        Icon(
-                            imageVector = favIcon,
-                            tint = Blue400,
-                            contentDescription = "isLike",
-                            modifier = Modifier.size(12.dp)
-                        )
-                        Spacer(modifier = Modifier.padding(2.dp))
-                        Text(item.likeCount.toString())
-                    }
+                    Text(
+                        text = "보여줄 항목이 없어요",
+                        style = Caption2,
+                        color = Gray600,
+                        modifier = Modifier.padding(start = 8.dp, top = 8.dp)
+                    )
+                }
+            }
+            LazyColumn {
+                items(
+                    count = myLikes.itemCount,
+                    key = { index -> myLikes[index]?.itemId ?: -1 }
+                ) { index ->
+                    val trade = myLikes[index] ?: return@items
+                    SummarizedTradeCard(
+                        item = trade,
+                        onItemCardClicked = {
+                            onItemClicked(trade.itemId)
+                        }
+                    )
                 }
             }
         }
-    }
-}
-
-@Composable
-private fun TradeItemStatus(isSold: Boolean) {
-    val backgroundColor = if (isSold) LightGray else Blue200
-    val text = if (isSold) "거래 완료" else "거래 가능"
-
-    Box(
-        modifier = Modifier
-            .clip(RoundedCornerShape(5.dp))
-            .background(backgroundColor),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(
-            text = text,
-            style = Body1,
-            color = White,
-            modifier = Modifier.padding(horizontal = 4.dp, vertical = 2.dp)
-        )
     }
 }
